Removed logo from dist and moved logo svgs to assets
